requirements.txt
Django==1.9
#mysql-python
djangorestframework
django-filter
django-crispy-forms
python-ldap
requests
python-dateutil